Releasing v1.0.20-devel
This release moves back to importing derivations from nixpkgs as source
code rather than relying on them in-place.

Occasionally the needs of nixpkgs drift from ours, and the changes in
nixpkgs can be difficult to reconcile with the needs of this project.
Therefore, the best solution right now is to just import the versions that
worked best with nixerl, and maintain them independently.

In addition, we've bumped the version of rebar3 used for Erlang versions
22 and above to 3.17, and made rebar3_16 available if folks want to use
an older version.