Rebuild Manifest
A new version of nix-prefetch-github has changed behaviour to use a
git-based fetch by default, because it fetches submodules. The
default behaviour of fetchGitHub, however, is to use the tarballs.

This means that some recently added versions have got the wrong hashes
computed for them (because the git sources, and the tarballs differ in
that the latter don't include .gitignore files).

This change tells the prefetcher not to get submodules, and also
rebuilds all metadata to ensure that no bad hashes are in the repo.